+++ files/CHANGELOG.md
@@ -6,13 +6,16 @@ Read more about it in `Rambda` changelog
 
 Also with this versions, typings tests are provided and several definitions are changed.
 
-2.16.0 Restore `R.runTests` with `src/extra/` folder
+2.16.0 Restore `R.runTests` but without documentation
+
+- export `getEvaluations`, `getPositiveEvaluation`, `getNegativeEvaluation` in the context of `R.runTests`
 
 2.15.0 Several changes
 
 - Typescript definitions have been updated and typings tests are introduced
 
 - `R.mapAsync` and `R.mapFastAsync` pass index as second argument
+
 
 2.14.1 Restore `R.contains`